+++ homework-5/homework-5/Delegates/SceneDelegate.swift
@@ -17,14 +17,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 		window = UIWindow(frame: windowScene.coordinateSpace.bounds)
 		window?.windowScene = windowScene
 
-		let navigationController = UINavigationController()
-		let assembly = AssemblyBuilder()
-		let router = Router(navigationController: navigationController, assembly: assembly)
-		router.initialViewController()
-		
+		let mainViewController = AssemblyMainBuilder.createMainModule()
+		let navigationController = UINavigationController(rootViewController: mainViewController)
 		window?.rootViewController = navigationController
 		window?.makeKeyAndVisible()
-
 	}
 }
