+++ homework-5/homework-5/Modules/DetailModule/View/DetailViewController.swift
@@ -7,6 +7,10 @@
 
 import UIKit
 
+protocol DetailViewProtocol: AnyObject {
+	func setComment(comment: Comments?)
+}
+
 final class DetailViewController: UIViewController {
 
 // MARK: - View
