--- homework-5/homework-5/Protocols/DetailProtocol/DetailViewPresenterProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  DetailViewPresenterProtocol.swift
-//  homework-5
-//
-//  Created by Денис Гынгазов on 12.11.2020.
-//
-
-import Foundation
-
-protocol DetailViewPresenterProtocol: AnyObject {
-	func viewComment()
-}
