--- homework-5/homework-5/Protocols/DetailProtocol/DetailViewProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  DetailViewProtocol.swift
-//  homework-5
-//
-//  Created by Денис Гынгазов on 12.11.2020.
-//
-
-import Foundation
-
-protocol DetailViewProtocol: AnyObject {
-	func setComment(comment: Comments?)
-}
