--- homework-5/homework-5/Protocols/MainProtocol/MainViewProtocol.swift
@@ -1,13 +1,0 @@
-//
-//  MainViewProtocol.swift
-//  homework-5
-//
-//  Created by Денис Гынгазов on 12.11.2020.
-//
-
-import Foundation
-
-protocol MainViewProtocol: AnyObject {
-	func succes()
-	func failure(error: Error)
-}
