--- homework-5/homework-5/Protocols/NetworksProtocol/NetworksProtocol.swift
@@ -1,12 +1,0 @@
-//
-//  NetworksProtocol.swift
-//  homework-5
-//
-//  Created by Денис Гынгазов on 12.11.2020.
-//
-
-import Foundation
-
-protocol NetworksProtocol {
-	func getComments(completion: @escaping (Result<[Comments]?, Error>) -> ())
-}
